47. Permutations II
